requirements.txt
fastapi==0.116.1
httpx==0.28.1
openai==1.108.1
pydantic-settings==2.11.0
pypdf==5.4.0
python-dotenv==1.1.1
sentence-transformers==5.1.1
torch>=2.6.0
uvicorn[standard]==0.35.0
